refactor: 작업 전 코드 정리
