+++ layer-api/src/main/java/org/layer/domain/retrospect/controller/RetrospectApi.java
@@ -1,6 +1,6 @@
-package org.layer.retrospect.controller;
+package org.layer.domain.retrospect.controller;
 
-import org.layer.retrospect.controller.dto.request.RetrospectCreateRequest;
+import org.layer.domain.retrospect.controller.dto.request.RetrospectCreateRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -12,7 +12,7 @@ import jakarta.validation.Valid;
 @Tag(name = "회고", description = "회고 관련 API")
 public interface RetrospectApi {
 
-	@Operation(summary = "회고 생성", description = "")
-	ResponseEntity<Void> createRetrospect(@PathVariable("spaceId") Long spaceId,
-		@RequestBody @Valid RetrospectCreateRequest request);
+    @Operation(summary = "회고 생성", description = "")
+    ResponseEntity<Void> createRetrospect(@PathVariable("spaceId") Long spaceId,
+                                          @RequestBody @Valid RetrospectCreateRequest request);
 }
