+++ layer-api/src/main/java/org/layer/domain/retrospect/controller/dto/request/RetrospectCreateRequest.java
@@ -1,4 +1,4 @@
-package org.layer.retrospect.controller.dto.request;
+package org.layer.domain.retrospect.controller.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
