+++ layer-common/src/main/java/org/layer/common/annotation/DisableSwaggerSecurity.java
@@ -1,0 +1,11 @@
+package org.layer.common.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface DisableSwaggerSecurity {
+}
